Calculate protein-protein interaction scores (PPI).

Using sequence counts before and after selection, calculate protein-protein interaction scores as described by Diss and Lehner (2018).

calculatePPIScore(
  se,
  pairingCol,
  ODCols,
  comparison,
  WTrows,
  selAssay = "counts"
)

Arguments

se

SummarizedExperiment object as returned by summarizeExperiment.

pairingCol

Name of column in colData(se) with replicate/pairing information. Samples with the same value in this column will be paired.

ODCols

Name(s) of column(s) in colData(se) with OD values (numeric), used to normalize for different numbers of cells.

comparison

3-element character vector of the form (column, numerator, denominator). column is the name of the column in colData(se) with experimental conditions. numerator and denominator define the comparison, e.g. c("cond", "output", "input") will look in the "cond" column and calculate PPI for the ratio of "output" over "input" counts.

WTrows

Vector of row names that will be used as the reference when calculating PPI scores. If more than one value is provided, the average of the corresponding PPI scores is used as a reference. If NULL, no division by WT scores will be done.

selAssay

Assay to select from se for the analysis.

Value

A numeric vector with PPI scores.

References

"The genetic landscape of a physical interaction." Diss G and Lehner B. Elife. 2018;7:e32472. doi: 10.7554/eLife.32472.

Author

Michael Stadler and Charlotte Soneson