Generate QC report

Generate QC report

Usage

generateQCReport(
  se,
  outFile,
  reportTitle = "mutscan QC report",
  forceOverwrite = FALSE,
  ...
)

Arguments

se

A SummarizedExperiment object, typically generated with summarizeExperiment().

outFile

Character string providing the name of the output file. Should have the extension .html.

reportTitle

Character string specifying the title of the QC report.

forceOverwrite

Logical scalar, indicating whether an existing file with the same name as outFile should be overwritten.

...

Additional parameters to be forwarded to render, for example quiet = TRUE.

Value

Invisibly, the path to the generated html file.

See also

Author

Charlotte Soneson

Examples

## Load SummarizedExperiment object
se <- readRDS(system.file("extdata", "GSE102901_cis_se.rds",
                          package = "mutscan"))
## Define output file
outfile <- tempfile(fileext = ".html")

## Generate QC report
generateQCReport(se, outfile)